initial refactor to separate component files
