--- src/component/BabyNamesList.js
@@ -1,9 +1,0 @@
-// Requirements
-// SUMMARY:
-// List of all baby names
-// NEEDS TO:
-// 1. Show the entire list on all device widths
-// 2. Show quick-glance information based on filters
-// 3. When a name is chosen for viewing (BabyNameDetails), the list needs to disappear and render the BabyNameDetails. It also needs to close this view with a one button click
-// DEPENDS ON:
-// * List of baby names from state
